Refactor Internal(Request|Response)Frame so they no longer inherit from Decoded(Request|Response)Frame
why:
1. InternalFrames are _not_ decoded
2. Need way know the encoded size of the message on arrival from the network.
   Existing inheritence hierarchy is impediment to that.


wip
