+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/frame/DecodedRequestFrame.java
@@ -5,20 +5,16 @@
  */
 package io.kroxylicious.proxy.frame;
 
-import org.apache.kafka.common.message.ProduceRequestData;
 import org.apache.kafka.common.message.RequestHeaderData;
 import org.apache.kafka.common.protocol.ApiMessage;
-
-import static org.apache.kafka.common.protocol.ApiKeys.PRODUCE;
 
 /**
  * A decoded request frame.
  */
 public class DecodedRequestFrame<B extends ApiMessage>
-        extends DecodedFrame<RequestHeaderData, B>
-        implements RequestFrame {
+        extends DecodedFrame<RequestHeaderData, B> implements ApiMessageBasedRequestFrame<B> {
 
-    private final boolean decodeResponse;
+    protected final boolean decodeResponse;
 
     public DecodedRequestFrame(short apiVersion,
                                int correlationId,
@@ -30,22 +26,8 @@ public class DecodedRequestFrame<B extends ApiMessage>
     }
 
     @Override
-    public short headerVersion() {
-        return apiKey().messageType.requestHeaderVersion(apiVersion);
-    }
-
-    @Override
     public boolean decodeResponse() {
         return decodeResponse;
-    }
-
-    @Override
-    public boolean hasResponse() {
-        return !isZeroAcksProduceRequest();
-    }
-
-    private boolean isZeroAcksProduceRequest() {
-        return apiKeyId() == PRODUCE.id && ((ProduceRequestData) body).acks() == 0;
     }
 
 }
