+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/internal/InternalResponseFrame.java
@@ -5,48 +5,22 @@
  */
 package io.kroxylicious.proxy.internal;
 
-import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
 
 import org.apache.kafka.common.message.ResponseHeaderData;
 import org.apache.kafka.common.protocol.ApiMessage;
 
 import io.kroxylicious.proxy.filter.Filter;
-import io.kroxylicious.proxy.frame.DecodedResponseFrame;
+import io.kroxylicious.proxy.frame.ApiMessageBasedResponseFrame;
 
-public class InternalResponseFrame<B extends ApiMessage> extends DecodedResponseFrame<B> {
-
-    private final Filter recipient;
-
-    private final CompletableFuture<?> future;
+public class InternalResponseFrame<B extends ApiMessage> extends InternalFrame<ResponseHeaderData, B> implements ApiMessageBasedResponseFrame<B> {
 
     public InternalResponseFrame(Filter recipient, short apiVersion, int correlationId, ResponseHeaderData header, B body, CompletableFuture<?> future) {
-        super(apiVersion, correlationId, header, body);
-        this.recipient = Objects.requireNonNull(recipient);
-        this.future = future;
+        super(apiVersion, correlationId, header, body, future, recipient);
     }
 
     public boolean isRecipient(Filter candidate) {
-        return recipient != null && recipient.equals(candidate);
+        return recipient() != null && recipient.equals(candidate);
     }
 
-    public Filter recipient() {
-        return recipient;
-    }
-
-    public CompletableFuture<?> promise() {
-        return future;
-    }
-
-    @Override
-    public String toString() {
-        return "InternalResponseFrame(" +
-                "recipient=" + recipient +
-                ", promise=" + future +
-                ", apiVersion=" + apiVersion +
-                ", correlationId=" + correlationId +
-                ", header=" + header +
-                ", body=" + body +
-                ')';
-    }
 }
